backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
reportlab==4.2.2
PyPDF2==3.0.1
gunicorn==22.0.0